Include statuscode in JsonPatchException
